Support to run cord-tester with a virtual environment where pip python packages are installed with prerequisites.sh Run prerequisites with: --venv option. Then use cord-tester bash script instead of using cord-test.py directly that wraps the invocation of cord-test.py by activating the virtual environment before running cord-tester.
In order to use cord-test.py directly as before with venv, just:
source setup/venv/bin/activate
before running cord-test.py.

On the cord or on CiaB, its recommended to install the pip packages on the host if one wants to use the MAAS/fabric test cases.
For cord/CiaB, just run the prerequisites with the --cord option as before.

Change-Id: I65e64e376540124165731da30a90152f559a5299